pyproject.toml
[project]
name = "biocoach-api"
version = "0.1.0"
description = "BioCoach — personal AI health advisor API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.6.0",
    "nats-py>=2.9.0",
    "openai>=1.55.0",
    "httpx>=0.28.0",
    "pyjwt[crypto]>=2.10.0",
    "python-telegram-bot>=21.7",
    "structlog>=24.4.0",
    "alembic>=1.14.0",
    "duckduckgo-search>=7.0.0",
    "pdfplumber>=0.11.0",
    "python-multipart>=0.0.18",
]

[tool.hatch.build.targets.wheel]
packages = ["api"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
